usrp1: Disallow device stopping with USRP1
GNU Radio libusrp driver hangs on internal control call
to 'set_rx_enable(false)' when stopping the device interface.
Because libusrp is no longer maintained, this is considered
a non-fixable case. As a result, revert USRP1 to previous
behaviour and disable attempts to stop the transceiver when
UHD is not used. Report failure as the corresponding response
on the socket interface.

Signed-off-by: Tom Tsou <tom@tsou.cc>